requirements.txt
fastapi
uvicorn
python-dotenv
pillow
opencv-python